requirements.txt
discord.py[voice]
python-dotenv
random2
pytube
requests
PyNaCl
openai
gtts
pydub